Add extra functionality from run.py to the trainer script (#856)
Fixes #339
