--- tests/test_run.py
@@ -1,14 +1,0 @@
-# -*- coding: utf-8 -*-
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this file,
-# You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import run
-
-
-def test_run():
-    # Test running the training for the bug model.
-    run.main(run.parse_args(["--train", "--goal", "defect"]))
-
-    # Test loading the trained model.
-    run.main(run.parse_args(["--goal", "defect"]))
